simulator/requirements.txt
google-cloud-pubsub==2.18.0
numpy==1.24.3
